dom/tests/mochitest/pointerlock/file_doubleLockCallBack.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=633602
-->
<head>
  <title>Bug 633602 - file_doubleLockCallBack.html</title>
  <script type="text/javascript"
          src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="text/javascript"
          src="/tests/SimpleTest/EventUtils.js"></script>
  <script type="application/javascript"
          src="mouselock_util.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
  <style type="text/css">
      #test-element { background-color: #94E01B; width:100px; height:100px; }
  </style>
</head>
<body onload="start();">
  <a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=633602">
    Mozilla Bug 633602</a>
  <div id="content">
    <div id="test-element"></div>
  </div>
  <pre id="test">
    <script type="text/javascript">
      /*
       * Test for Bug 633602
       * This test checks to ensure that if pointer.lock is called on the same
       * element twice, it still fires the successful lock callback both times.
       */
      SimpleTest.waitForExplicitFinish();
      var pointer = navigator.mozPointer;
      var testElement = document.getElementById("test-element");

      document.addEventListener("mozfullscreenchange", function (e) {
        if (document.mozFullScreen &&
            document.mozFullScreenElement === testElement) {
          pointer.lock(testElement, function () {
            // Call lock again on 1st Success callback
            pointer.lock(testElement, function() {
              ok(pointer.isLocked, "Mouse was locked before this call, " +
                "Success Callback of nested call executed again");
              document.mozCancelFullScreen();
              SimpleTest.finish();
            }, function() {
              ok(false, "Mouse failed to double lock same element");
              SimpleTest.finish();
            });
          }, function () {
            ok(pointer.isLocked, "Mouse failed to lock fullscreened element");
            SimpleTest.finish();
          });
        }
      }, false);

      function start() {
        SimpleTest.waitForFocus(function() {
          testElement.mozRequestFullScreen();
        });
      }
    </script>
  </pre>
</body>
</html>
